blacksmithAI/pyproject.toml
[project]
name = "blacksmithAI"
version = "1.0.0"
description = "AI Powered Penetration Testing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.0",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.0",
    "langchain-chroma>=0.1.2",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "langgraph-cli[inmem]>=0.4.11",
    "rich>=14.2.0",
]
